Drasil.DocDecl

Description

Document declaration types and functions for generating Software Requirement Specifications.

Types

type SRSDecl = [DocSection]

A Software Requirements Specification Declaration is made up of all necessary sections (DocSections).

data DocSection

Contains all the different sections needed for a full SRS (SRSDecl).

Constructors

TableOfContents
Table of Contents

RefSec RefSec
Reference.

IntroSec IntroSec
Introduction.

StkhldrSec StkhldrSec
Stakeholders.

GSDSec GSDSec
General System Description.

SSDSec SSDSec
Specific System Description.

ReqrmntSec ReqrmntSec
Requirements.

LCsSec
Likely Changes.

UCsSec
Unlikely Changes.

TraceabilitySec TraceabilitySec
Traceability.

AuxConstntSec AuxConstntSec
Auxiliary Constants.

Bibliography
Bibliography.

AppndxSec AppndxSec
Appendix.

OffShelfSolnsSec OffShelfSolnsSec
Off the Shelf Solutions.

newtype SSDSec

Specific System Description section (wraps SSDSub subsections).

Constructors

SSDProg [SSDSub]

data SSDSub where

Specific System Description subsections.

Constructors

SSDProblem :: ProblemDescription -> SSDSub
Problem description.

SSDSolChSpec :: SolChSpec -> SSDSub
Solution characteristics.

data ProblemDescription where

Problem Description section.

Constructors

PDProg :: Sentence -> [Section] -> [PDSub] -> ProblemDescription

data PDSub where

Problem Description subsections.

Constructors

TermsAndDefs :: Concept c => Maybe Sentence -> [c] -> PDSub
Terms and Definitions.

PhySysDesc :: Idea a => a -> [Sentence] -> LabelledContent -> [Contents] -> PDSub
Physical System Description.

Goals :: [Sentence] -> PDSub
Goals.

data SolChSpec where

Solution Characteristics Specification section (wraps SCSSub subsections).

Constructors

SCSProg :: [SCSSub] -> SolChSpec

data SCSSub where

Solution Characteristics Specification subsections.

Constructors

Assumptions :: SCSSub
Assumptions.

TMs :: [Sentence] -> Fields -> SCSSub
Theory models.

GDs :: [Sentence] -> Fields -> DerivationDisplay -> SCSSub
General definitions.

DDs :: [Sentence] -> Fields -> DerivationDisplay -> SCSSub
Data definitions.

IMs :: [Sentence] -> Fields -> DerivationDisplay -> SCSSub
Instance models.

Constraints :: (HasUncertainty c, Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => Sentence -> [c] -> SCSSub
Constraints.

CorrSolnPpties :: (Quantity c, Constrained c) => [c] -> [Contents] -> SCSSub
Properties of a correct solution.

newtype ReqrmntSec

Requirements section (wraps ReqsSub subsections).

Constructors

ReqsProg [ReqsSub]

data ReqsSub where

Requirements subsections.

Constructors

FReqsSub :: Sentence -> [LabelledContent] -> ReqsSub
Functional requirements. LabelledContent for tables (includes input values).

FReqsSub' :: [LabelledContent] -> ReqsSub
Functional requirements. LabelledContent for tables (no input values).

NonFReqsSub :: ReqsSub
Non-Functional requirements.

Functions

mkDocDesc :: System -> SRSDecl -> DocDesc

Creates the document description (translates SRSDecl into a more usable form for generating documents).