Drasil.DocLang

Documentation

type SRSDecl = [DocSection]

data DocSection

Constructors

RefSec RefSec

IntroSec IntroSec

StkhldrSec StkhldrSec

GSDSec GSDSec

SSDSec SSDSec

ReqrmntSec ReqrmntSec

LCsSec

UCsSec

TraceabilitySec TraceabilitySec

AuxConstntSec AuxConstntSec

Bibliography

AppndxSec AppndxSec

OffShelfSolnsSec OffShelfSolnsSec

newtype ReqrmntSec

Constructors

ReqsProg [ReqsSub]

data ReqsSub where

Constructors

FReqsSub :: Sentence -> [LabelledContent] -> ReqsSub

FReqsSub' :: [LabelledContent] -> ReqsSub

NonFReqsSub :: ReqsSub

data PDSub where

Problem Description subsections

Constructors

TermsAndDefs :: Concept c => Maybe Sentence -> [c] -> PDSub

PhySysDesc :: Idea a => a -> [Sentence] -> LabelledContent -> [Contents] -> PDSub

Goals :: [Sentence] -> PDSub

data ProblemDescription where

Problem Description section

Constructors

PDProg :: Sentence -> [Section] -> [PDSub] -> ProblemDescription

newtype SSDSec

Specific System Description section . Contains a list of subsections.

Constructors

SSDProg [SSDSub]

data SSDSub where

Specific system description subsections

Constructors

SSDProblem :: ProblemDescription -> SSDSub

SSDSolChSpec :: SolChSpec -> SSDSub

data SCSSub where

Solution Characteristics Specification subsections

Constructors

Assumptions :: SCSSub

TMs :: [Sentence] -> Fields -> SCSSub

GDs :: [Sentence] -> Fields -> DerivationDisplay -> SCSSub

DDs :: [Sentence] -> Fields -> DerivationDisplay -> SCSSub

IMs :: [Sentence] -> Fields -> DerivationDisplay -> SCSSub

Constraints :: (HasUncertainty c, Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => Sentence -> [c] -> SCSSub

CorrSolnPpties :: (Quantity c, Constrained c) => [c] -> [Contents] -> SCSSub

data SolChSpec where

Solution Characteristics Specification section

Constructors

SCSProg :: [SCSSub] -> SolChSpec

mkDoc :: SRSDecl -> (IdeaDict -> IdeaDict -> Sentence) -> SystemInformation -> Document

Creates a document from a document description and system information

tsymb :: [TSIntro] -> RefTab

Default Term and given intro

table of symbols constructor

tsymb'' :: [TSIntro] -> LFunc -> RefTab

Custom function and intro.

Custom table of symbols constructor

tunit :: [TUIntro] -> RefTab

table of units constructors

tunit' :: [TUIntro] -> RefTab

table of units constructors

newtype AppndxSec

Constructors

AppndxProg [Contents]

data AuxConstntSec

Values of Auxiliary Constants section

Constructors

AuxConsProg CI [QDefinition]

data DerivationDisplay

Constructors

ShowDerivation

HideDerivation

data Emphasis

How to handle Emphasis

Constructors

Bold

Italics

newtype OffShelfSolnsSec

Off-The-Shelf Solutions section

Constructors

OffShelfSolnsProg [Contents]

newtype GSDSec

Constructors

GSDProg [GSDSub]

data GSDSub where

Constructors

SysCntxt :: [Contents] -> GSDSub

UsrChars :: [Contents] -> GSDSub

SystCons :: [Contents] -> [Section] -> GSDSub

data IntroSec

Introduction section. Contents are top level followed by a list of subsections.

Constructors

IntroProg Sentence Sentence [IntroSub]
Temporary, will be modified once we've figured out more about the section.

data IntroSub where

Introduction subsections

Constructors

IPurpose :: [Sentence] -> IntroSub

IScope :: Sentence -> IntroSub

IChar :: [Sentence] -> [Sentence] -> [Sentence] -> IntroSub

IOrgSec :: Sentence -> CI -> Section -> Sentence -> IntroSub

data LFunc where

Lens (lookup) functions (currently for TSymb)

Constructors

Term :: LFunc

Defn :: LFunc

TermExcept :: [DefinedQuantityDict] -> LFunc

DefnExcept :: [DefinedQuantityDict] -> LFunc

TAD :: LFunc

data Literature

Types of literature

Constructors

Lit Topic
literature

Doc' Topic
existing documentation for (plural of topic)

Manual Topic
manual

data RefSec

Reference section. Contents are top level followed by a list of subsections.

Constructors

RefProg Contents [RefTab]

data RefTab where

Reference subsections

Constructors

TUnits :: RefTab

TUnits' :: [TUIntro] -> ([UnitDefn] -> LabelledContent) -> RefTab

TSymb :: [TSIntro] -> RefTab

TSymb' :: LFunc -> [TSIntro] -> RefTab

TAandA :: RefTab

newtype StkhldrSec

Stakeholders section

Constructors

StkhldrProg [StkhldrSub]

data StkhldrSub where

Stakeholders subsections

Constructors

Client :: CI -> Sentence -> StkhldrSub

Cstmr :: CI -> StkhldrSub

data TConvention

Possible typographic conventions

Constructors

Vector Emphasis
How vectors are emphasized

Verb Sentence
Verbatim for specialized conventions

newtype TraceabilitySec

Constructors

TraceabilityProg [TraceConfig]

data TSIntro

For creating the table of symbols intro

Constructors

TypogConvention [TConvention]
Typographic conventions used

SymbOrder
Symbol ordering (defaults to alphabetical)

SymbConvention [Literature]
Symbol conventions match specified literature

TSPurpose
Purpose of the Table of Symbols

VectorUnits
Definition of vector components

data TUIntro

For creating the table of units intro

Constructors

System
System of units (defaults to SI)

Derived
Sentence about derived units being used alongside SI

TUPurpose
Purpose of the table of units

data Field

Fields that should be displayed in definitions

Constructors

Label

Symbol

Units

DefiningEquation

Description Verbosity InclUnits

Input

Output

InConstraints

OutConstraints

Notes

Source

RefBy

type Fields = [Field]

Synonym for a list of Field

data InclUnits

Constructors

IncludeUnits

data Verbosity

Constructors

Verbose

Succinct

ddefn :: Fields -> SystemInformation -> DataDefinition -> LabelledContent

Create a data definition using a list of fields, a database of symbols, and a QDefinition (called automatically by SCSSub program)

data ModelDB

ddRefDB :: ModelDB -> RefMap DataDefinition

mdb :: [TheoryModel] -> [GenDefn] -> [DataDefinition] -> [InstanceModel] -> ModelDB

purpDoc :: CI -> Verbosity -> [Sentence]

intro :: Contents

Default reference section introduction

inReq :: Sentence -> ConceptInstance

inTable :: (Quantity i, MayHaveUnit i) => [i] -> LabelledContent

mkInputPropsTable :: (Quantity i, MayHaveUnit i, HasShortName r, Referable r) => [i] -> r -> LabelledContent

takes a list of wrapped variables and creates an Input Data Table for uses in Functional Requirments

mkQRTuple :: (Quantity i, MayHaveUnit i, HasShortName i, Referable i) => [i] -> [(QuantityDict, Sentence)]

mkQRTupleRef :: (Quantity i, MayHaveUnit i, HasShortName r, Referable r) => [i] -> [r] -> [(QuantityDict, Sentence)]

mkValsSourceTable :: (Quantity i, MayHaveUnit i) => [(i, Sentence)] -> String -> Sentence -> LabelledContent

takes a list of tuples of variables and sources and creates an table for uses in Functional Requirments

auxSpecSent :: Sentence

termDefnF' :: Maybe Sentence -> [Contents] -> Section

traceMatStandard :: SystemInformation -> [TraceConfig]

getDocDesc :: DocDesc -> [Sentence]

egetDocDesc :: DocDesc -> [Expr]

ciGetDocDesc :: DocDesc -> [CI]

generateTraceMap :: [DocSection] -> TraceMap

solutionLabel :: Reference

characteristicsLabel :: Reference