Drasil.DocLang.Notebook

Documentation

intro :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

prpsOfDoc :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

body :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

review :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

mainIdea :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

motion :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

hormotion :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

vermotion :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

methAndAnls :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

coorSyst :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

kinematic :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

procForAnls :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

summary :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

appendix :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

reference :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

example :: [Contents] -> [Section] -> Section

Notebook constructor. Create the notebook from given system name, authors, and sections doc :: NamedIdea c => c -> Sentence -> [Section] -> Document doc sys = Document (Doc.notebook forTPS sys)

sectionReferences :: [Reference]

introLabel :: Reference

docPurposeLabel :: Reference

methsAndanlsLabel :: Reference

referenceLabel :: Reference

bodyLabel :: Reference

reviewLabel :: Reference

mainIdeaLabel :: Reference

motionLabel :: Reference

hormotionLabel :: Reference

vermotionLabel :: Reference

appendixLabel :: Reference

coorSystLabel :: Reference

kinematicLabel :: Reference

summaryLabel :: Reference

anlsProcLabel :: Reference

exampleLabel :: Reference