Drasil.DocumentLanguage

Description

Start the process of moving away from Document as the main internal representation of information, to something more informative. Over time, we'll want to have a cleaner separation, but doing that all at once would break too much for too long. So we start here instead.

Documentation

mkDoc :: SRSDecl -> (IdeaDict -> IdeaDict -> Sentence) -> SystemInformation -> Document

Creates a document from a document description and system information

extractUnits :: DocDesc -> ChunkDB -> [UnitDefn]

fillTraceMaps :: DocDesc -> SystemInformation -> SystemInformation

fillReqs :: DocDesc -> SystemInformation -> SystemInformation

mkSections :: SystemInformation -> DocDesc -> [Section]

Helper for creating the document sections

mkRefSec :: SystemInformation -> DocDesc -> RefSec -> Section

Helper for creating the reference section and subsections

tunit :: [TUIntro] -> RefTab

table of units constructors

tunit' :: [TUIntro] -> RefTab

table of units constructors

mkTSymb :: (Quantity e, Concept e, Eq e, MayHaveUnit e) => [e] -> LFunc -> [TSIntro] -> Section

Helper for creating the table of symbols

tsymb :: [TSIntro] -> RefTab

Default Term and given intro

table of symbols constructor

tsymb' :: [TSIntro] -> RefTab

Default Defn and given intro

table of symbols constructor

tsymb'' :: [TSIntro] -> LFunc -> RefTab

Custom function and intro.

Custom table of symbols constructor

tsIntro :: [TSIntro] -> Contents

table of symbols intro builder. Used by mkRefSec

tsI :: TSIntro -> Sentence

table of symbols intro writer. Translates a TSIntro to a list of Sentences

typogConvention :: [TConvention] -> Sentence

typographic convention writer. Translates a list of typographic conventions to a sentence

symbConvention :: [Literature] -> Sentence

symbolic convention writer.

tuIntro :: [TUIntro] -> Contents

Table of units intro builder. Used by mkRefSec

tuI :: TUIntro -> Sentence

table of units intro writer. Translates a TUIntro to a Sentence.

defaultTUI :: [TUIntro]

Default table of units intro contains the

mkIntroSec :: SystemInformation -> IntroSec -> Section

mkStkhldrSec :: StkhldrSec -> Section

Helper for making the Stakeholders section

mkGSDSec :: GSDSec -> Section

Helper for making the 'General System Description' section

mkSSDSec :: SystemInformation -> SSDSec -> Section

Helper for making the 'Specific System Description' section

mkSSDProb :: SystemInformation -> ProblemDescription -> Section

mkSolChSpec :: SystemInformation -> SolChSpec -> Section

helperCI :: ConceptInstance -> SystemInformation -> ConceptInstance

tmStub :: Section

Section stubs for implicit referencing

ddStub :: Section

Section stubs for implicit referencing

imStub :: Section

Section stubs for implicit referencing

pdStub :: Section

Section stubs for implicit referencing

mkReqrmntSec :: ReqrmntSec -> Section

Helper for making the Requirements section

mkLCsSec :: LCsSec -> Section

Helper for making the LikelyChanges section

mkUCsSec :: UCsSec -> Section

Helper for making the UnikelyChanges section

mkTraceabilitySec :: TraceabilitySec -> SystemInformation -> Section

Helper for making the 'Traceability Matrices and Graphs' section

mkOffShelfSolnSec :: OffShelfSolnsSec -> Section

Helper for making the 'Off-the-Shelf Solutions' section

mkAuxConsSec :: AuxConstntSec -> Section

Helper for making the 'Values of Auxiliary Constants' section

mkBib :: BibRef -> Section

Helper for making the bibliography section

mkAppndxSec :: AppndxSec -> Section

Helper for making the Appendix section

siSys :: SystemInformation -> IdeaDict