Drasil.DocumentLanguage.Core

Description

Defines core types for use with the Drasil document language (Drasil.DocumentLanguage).

Documentation

type System = Sentence

Type synonym for clarity.

type DocKind = Sentence

Type synonym for clarity.

Document Types

type DocDesc = [DocSection]

A document description is made up of document sections.

data DocSection

Document sections are either Reference, Introduction, or Specific System Description sections (for now!).

Constructors

TableOfContents

RefSec RefSec

IntroSec IntroSec

StkhldrSec StkhldrSec

GSDSec GSDSec

SSDSec SSDSec

ReqrmntSec ReqrmntSec

LCsSec LCsSec

UCsSec UCsSec

TraceabilitySec TraceabilitySec

AuxConstntSec AuxConstntSec

Bibliography

AppndxSec AppndxSec

OffShelfSolnsSec OffShelfSolnsSec

Reference Material Section

data RefSec

Reference section. Contents are top level followed by a list of subsections.

Constructors

RefProg Contents [RefTab]

data RefTab where

Reference subsections (tables) made out of units or symbols (can be customized).

Constructors

TUnits :: RefTab
Default table of units.

TUnits' :: [TUIntro] -> ([UnitDefn] -> LabelledContent) -> RefTab
Customized introduction.

TSymb :: [TSIntro] -> RefTab
Adds an introduction for a table of symbols.

TSymb' :: LFunc -> [TSIntro] -> RefTab
Allows Lens functions in addition to an introduction for a table of symbols.

TAandA :: [IdeaDict] -> RefTab
Default.

data TSIntro

For creating a table of symbols introduction

Constructors

TypogConvention [TConvention]
Typographic conventions used.

SymbOrder
Symbol ordering (defaults to alphabetical).

SymbConvention [Literature]
Symbol conventions match specified literature.

TSPurpose
Purpose of the Table of Symbols.

VectorUnits
Definition of vector components.

data TConvention

Possible typographic conventions.

Constructors

Vector Emphasis
How vectors are emphasized.

Verb Sentence
Verbatim for specialized conventions.

data Emphasis

How to handle emphasis of words.

Constructors

Bold

Italics

data Literature

Types of literature.

Constructors

Lit Topic
Literature (with a Topic).

Doc Topic
Existing documentation for (singular topic).

Doc' Topic
Existing documentation for (plural version of topic).

Manual Topic
Manual.

type Topic = IdeaDict

Type synonym for clarity.

data TUIntro

For creating the table of units introduction.

Constructors

System
System of units (defaults to SI).

Derived
Sentence about derived units being used alongside SI.

TUPurpose
Purpose of the table of units.

data LFunc where

Lens (lookup) functions (currently for TSymb).

Constructors

Term :: LFunc

Defn :: LFunc

TermExcept :: [DefinedQuantityDict] -> LFunc

DefnExcept :: [DefinedQuantityDict] -> LFunc

TAD

Fields

:: LFunc

Term and Definition.

Introduction Section

data IntroSec

Introduction section. Contents are top level followed by a list of subsections.

Constructors

IntroProg Sentence Sentence [IntroSub]
Temporary, will be modified once we've figured out more about the section.

data IntroSub where

Introduction subsections.

Constructors

IPurpose :: [Sentence] -> IntroSub
Describes purpose of the system.

IScope :: Sentence -> IntroSub
Describes scope of the system.

IChar :: [Sentence] -> [Sentence] -> [Sentence] -> IntroSub
Describes characteristics of the system.

IOrgSec :: CI -> Section -> Sentence -> IntroSub
Organises the section.

Stakeholders Section

newtype StkhldrSec

Stakeholders section (wraps stakeholders subsections StkhldrSub).

Constructors

StkhldrProg [StkhldrSub]

data StkhldrSub where

Stakeholders subsections.

Constructors

Client :: CI -> Sentence -> StkhldrSub
May have a client.

Cstmr :: CI -> StkhldrSub
May have a customer.

General System Description Section

newtype GSDSec

General System Description section (wraps GSDSub subsections).

Constructors

GSDProg [GSDSub]

data GSDSub where

General System Description subsections.

Constructors

SysCntxt :: [Contents] -> GSDSub
System context.

UsrChars :: [Contents] -> GSDSub
User characteristics.

SystCons :: [Contents] -> [Section] -> GSDSub
System constraints.

Specific System Description Section

newtype SSDSec

Specific System Description section. Contains a list of subsections (SSDSub).

Constructors

SSDProg [SSDSub]

data SSDSub where

Specific System Description subsections.

Constructors

SSDProblem :: ProblemDescription -> SSDSub
System description problems.

SSDSolChSpec :: SolChSpec -> SSDSub
Solution characteristics specification.

data ProblemDescription where

Problem Description section. Contains an intro or title, Sections, and problem description subsections (PDSub).

Constructors

PDProg :: Sentence -> [Section] -> [PDSub] -> ProblemDescription

data PDSub where

Problem Description subsections.

Constructors

TermsAndDefs :: Concept c => Maybe Sentence -> [c] -> PDSub
Terms and definitions.

PhySysDesc :: Idea a => a -> [Sentence] -> LabelledContent -> [Contents] -> PDSub
Physical system description.

Goals :: [Sentence] -> [ConceptInstance] -> PDSub
Goals.

data SolChSpec where

Solution Characteristics Specification section. Contains a list of subsections (SCSSub).

Constructors

SCSProg :: [SCSSub] -> SolChSpec

data SCSSub where

Solution Characteristics Specification subsections.

Constructors

Assumptions :: [ConceptInstance] -> SCSSub
Assumptions.

TMs :: [Sentence] -> Fields -> [TheoryModel] -> SCSSub
Theory Models.

GDs :: [Sentence] -> Fields -> [GenDefn] -> DerivationDisplay -> SCSSub
General Definitions.

DDs :: [Sentence] -> Fields -> [DataDefinition] -> DerivationDisplay -> SCSSub
Data Definitions.

IMs :: [Sentence] -> Fields -> [InstanceModel] -> DerivationDisplay -> SCSSub
Instance Models.

Constraints :: (HasUncertainty c, Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => Sentence -> [c] -> SCSSub
Constraints.

CorrSolnPpties :: (Quantity c, Constrained c) => [c] -> [Contents] -> SCSSub
Properties of a correct solution.

data DerivationDisplay

Choose whether to show or hide the derivation of an expression.

Constructors

ShowDerivation

HideDerivation

Requirements Section

newtype ReqrmntSec

Requirements section. Contains a list of subsections (ReqsSub).

Constructors

ReqsProg [ReqsSub]

data ReqsSub where

Requirements subsections.

Constructors

FReqsSub' :: [ConceptInstance] -> [LabelledContent] -> ReqsSub
Functional requirements. LabelledContent needed for tables.

FReqsSub :: [ConceptInstance] -> [LabelledContent] -> ReqsSub
Functional requirements. LabelledContent needed for tables.

NonFReqsSub :: [ConceptInstance] -> ReqsSub
Non-functional requirements.

Likely Changes Section

newtype LCsSec

Likely Changes section.

Constructors

LCsProg [ConceptInstance]

Unlikely Changes Section

newtype UCsSec

Unlikely Changes section.

Constructors

UCsProg [ConceptInstance]

Traceability Section

newtype TraceabilitySec

Traceability Matices and Graphs section. Contains configurations (TraceConfig).

Constructors

TraceabilityProg [TraceConfig]

data TraceConfig

Traceability Matices and Graphs configurations.

Constructors

TraceConfig UID [Sentence] Sentence [TraceViewCat] [TraceViewCat]

getTraceConfigUID :: TraceConfig -> UID

Off-The-Shelf Solutions Section

newtype OffShelfSolnsSec

Off-The-Shelf Solutions section.

Constructors

OffShelfSolnsProg [Contents]

Values of Auxiliary Constants Section

data AuxConstntSec

Values of Auxiliary Constants section.

Constructors

AuxConsProg CI [ConstQDef]

Appendix Section

newtype AppndxSec

Appendix section.

Constructors

AppndxProg [Contents]

Multiplate Definition and Type

data DLPlate f

Holds all of the different kinds of sections. Defines as a plate with an applicative functor.

Constructors

DLPlate

Fields

docSec :: DocSection -> f DocSection

refSec :: RefSec -> f RefSec

introSec :: IntroSec -> f IntroSec

introSub :: IntroSub -> f IntroSub

stkSec :: StkhldrSec -> f StkhldrSec

stkSub :: StkhldrSub -> f StkhldrSub

gsdSec :: GSDSec -> f GSDSec

gsdSub :: GSDSub -> f GSDSub

ssdSec :: SSDSec -> f SSDSec

ssdSub :: SSDSub -> f SSDSub

pdSec :: ProblemDescription -> f ProblemDescription

pdSub :: PDSub -> f PDSub

scsSub :: SCSSub -> f SCSSub

reqSec :: ReqrmntSec -> f ReqrmntSec

reqSub :: ReqsSub -> f ReqsSub

lcsSec :: LCsSec -> f LCsSec

ucsSec :: UCsSec -> f UCsSec

traceSec :: TraceabilitySec -> f TraceabilitySec

offShelfSec :: OffShelfSolnsSec -> f OffShelfSolnsSec

auxConsSec :: AuxConstntSec -> f AuxConstntSec

appendSec :: AppndxSec -> f AppndxSec