Drasil.DocumentLanguage.Definitions

Description

Defines helper functions for creating subsections within the Solution Characteristics Specification. Namely, for theory models, general definitions, data definitions, and instance models.

Types

data Field

Fields that should be displayed in definitions.

Constructors

Label

Symbol

Units

DefiningEquation

Description Verbosity InclUnits

Input

Output

InConstraints

OutConstraints

Notes

Source

RefBy

type Fields = [Field]

Synonym for a list of Fields.

data InclUnits

Determines whether to include or ignore units.

Constructors

IncludeUnits
In description field (for other symbols).

IgnoreUnits

data Verbosity

Refers to the verbosity of statements.

Constructors

Verbose
Full Descriptions.

Succinct
Simple Description (do not redefine other symbols).

Constructors

ddefn :: Fields -> System -> DataDefinition -> LabelledContent

Create a data definition using a list of fields, a database of symbols, and a QDefinition (called automatically by SCSSub program).

derivation :: (MayHaveDerivation c, HasShortName c, Referable c) => c -> Maybe Contents

Create a derivation from a chunk's attributes. This follows the TM, DD, GD, or IM definition automatically (called automatically by SCSSub program).

gdefn :: Fields -> System -> GenDefn -> LabelledContent

Create a general definition using a list of fields, database of symbols, and a GenDefn (general definition) chunk (called automatically by SCSSub program).

instanceModel :: Fields -> System -> InstanceModel -> LabelledContent

Create an instance model using a list of fields, database of symbols, and an InstanceModel chunk (called automatically by SCSSub program).

tmodel :: Fields -> System -> TheoryModel -> LabelledContent

Create a theoretical model using a list of fields to be displayed, a database of symbols, and a RelationConcept (called automatically by SCSSub program).

Helpers

helperRefs :: HasUID t => t -> System -> Sentence

Helper function to make a list of Sentences from the current system information and something that has a UID.

helpToRefField :: UID -> System -> Sentence

Creates a reference as a Sentence by finding if the UID is in one of the possible data sets contained in the System database.