Drasil.DocumentLanguage.Notebook.LsnDecl

Description

Lesson plan notebook section declaration types and functions.

Types

type LsnDecl = [LsnChapter]

A Lesson Plan notebook declaration is made up of all necessary chapters (LsnChapters).

data LsnChapter

Contains all the different chapters needed for a notebook lesson plan (LsnDecl).

Constructors

Intro Intro

LearnObj LearnObj

Review Review

CaseProb CaseProb

Example Example

Smmry Smmry

BibSec

Apndx Apndx

Functions

mkLsnDesc :: System -> LsnDecl -> LsnDesc

Creates the lesson description (translates LsnDecl into a more usable form for generating documents).