Drasil.DocumentLanguage.Notebook.NBDecl

Description

Lesson plan notebook section declaration types and functions.

Types

type NBDecl = [NbSection]

A Lesson Plan notebook declaration is made up of all necessary sections (NbSections).

data NbSection

Contains all the different sections needed for a notebook lesson plan (NBDecl).

Constructors

IntrodSec IntrodSec

BodySec BodySec

SmmrySec SmmrySec

BibSec

ApndxSec ApndxSec

Functions

mkNBDesc :: SystemInformation -> NBDecl -> NBDesc

Creates the notebook description (translates NBDecl into a more usable form for generating documents).