Drasil.DocumentLanguage.TraceabilityGraph

Description

Defines functions to create traceability graphs in SRS documents.

Documentation

traceMGF :: [LabelledContent] -> [Sentence] -> [Contents] -> String -> [Section] -> Section

Wrapper for traceMIntro and traceGIntro. Turns references (LabelledContents), trailing notes (Sentences), and any other needed contents to create a Traceability Section. Traceability graphs generate as both a link and a figure for convenience.

traceyGraphGetRefs :: String -> [Reference]

Create references to traceability graphs given an example name. Primarily used for reference database in examples.

genTraceGraphLabCons :: String -> [LabelledContent]

Generate the LabelledContent chunks, specialized to a specific example (and its path).

mkGraphInfo :: System -> GraphInfo

Extracts traceability graph inforomation from filled-in System.

resourcePath :: String

Traceability graphs reference path.