Drasil.ExtractLsnDesc

Description

Defines functions to extract citation references from Notebook documents.

Documentation

extractLsnPlanBib :: System -> LsnDesc -> BibRef

Extract bibliography entries for a notebook based on the lesson description. Scans the notebook for citation references and looks them up in the database.