Drasil.GetChunks

Description

Utilities to get grab certain chunks (from Expr, Sentence, etc) by UID and dereference the chunk it refers to.

Documentation

ccss :: [Sentence] -> [ModelExpr] -> ChunkDB -> [DefinedQuantityDict]

Gets a list of defined quantities (DefinedQuantityDicts) from Sentences and expressions that are contained in the database (ChunkDB).

ccss' :: [Sentence] -> [ModelExpr] -> ChunkDB -> [DefinedQuantityDict]

Gets a list of quantities (DefinedQuantityDicts) from Sentences and expressions that are contained in the database (ChunkDB).

combine :: Sentence -> ChunkDB -> [DefinedQuantityDict]

Combines the functions of vars and concpt to create a list of DefinedQuantityDicts from a Sentence.

vars :: ModelExpr -> ChunkDB -> [DefinedQuantityDict]

Gets a list of quantities (DefinedQuantityDict) from an equation in order to print.

citeDB :: System -> BibRef

Extract bibliography entries for a system.