Drasil.SRSDocument

Description

Re-export functions common to all SRS-related examples. This aims to reduce clutter at the top of every Body.hs file in the Drasil examples. If an import is changed here, it will be changed for every example.

Chunk database types and functions

Imported from Database.Drasil

data ChunkDB

Our chunk databases. Must contain all maps needed in an example. In turn, these maps must contain every chunk definition or concept used in its respective example, else an error is thrown.

data System where

Data structure for holding all of the requisite information about a system to be used in artifact generation.

Constructors

SI

Fields

:: forall a e h i j. (CommonIdea a, Idea a, Quantity e, Eq e, MayHaveUnit e, Concept e, Quantity h, MayHaveUnit h, Concept h, Quantity i, MayHaveUnit i, Concept i, HasUID j, Constrained j)

=> { _sys :: a

, _kind :: SystemKind

, _authors :: People

, _purpose :: Purpose

, _background :: Background

, _scope :: Scope

, _motivation :: Motivation

, _quants :: [e]

, _theoryModels :: [TheoryModel]

, _genDefns :: [GenDefn]

, _dataDefns :: [DataDefinition]

, _instModels :: [InstanceModel]

, _configFiles :: [String]

, _inputs :: [h]

, _outputs :: [i]

, _constraints :: [j]

, _constants :: [ConstQDef]

, _systemdb :: ChunkDB

} -> System

cdb' :: (Quantity q, MayHaveUnit q, Concept q, Concept c, IsUnit u) => [q] -> [IdeaDict] -> [c] -> [u] -> [DataDefinition] -> [InstanceModel] -> [GenDefn] -> [TheoryModel] -> [ConceptInstance] -> [LabelledContent] -> [Reference] -> [Citation] -> ChunkDB

Smart constructor for chunk databases. Takes in the following:

[Quantity] (for SymbolMap),

NamedIdeas (for TermMap),

Concepts (for ConceptMap),

Units (something that IsUnit for UnitMap),

DataDefinitions (for DatadefnMap),

InstanceModels (for InsModelMap),

GenDefns (for GendefMap),

TheoryModels (for TheoryModelMap),

ConceptInstances (for ConceptInstanceMap),

LabelledContents (for LabelledContentMap). Creates an empty ChunkDB without basic data. cdb in MetaDatabase.Drasil should be preferred over this function as it includes the basic data needed for Drasil.

Printing Information needed to generate all documents

Imported from Language.Drasil.Printers

data PrintingInformation

Printing information contains a database, a stage, and a printing configuration.

Constructors

PI

Fields

_ckdb :: ChunkDB

_stg :: Stage

_configuration :: PrintingConfiguration

defaultConfiguration :: PrintingConfiguration

Default configuration is for engineering.

piSys :: System -> Stage -> PrintingConfiguration -> PrintingInformation

Builds a document's printing information based on the system information.

Document section types needed for a SRS

Imported from Drasil.DocDecl

type SRSDecl = [DocSection]

A Software Requirements Specification Declaration is made up of all necessary sections (DocSections).

data DocSection

Contains all the different sections needed for a full SRS (SRSDecl).

Constructors

TableOfContents
Table of Contents

RefSec RefSec
Reference.

IntroSec IntroSec
Introduction.

StkhldrSec StkhldrSec
Stakeholders.

GSDSec GSDSec
General System Description.

SSDSec SSDSec
Specific System Description.

ReqrmntSec ReqrmntSec
Requirements.

LCsSec
Likely Changes.

UCsSec
Unlikely Changes.

TraceabilitySec TraceabilitySec
Traceability.

AuxConstntSec AuxConstntSec
Auxiliary Constants.

Bibliography
Bibliography.

AppndxSec AppndxSec
Appendix.

OffShelfSolnsSec OffShelfSolnsSec
Off the Shelf Solutions.

newtype ReqrmntSec

Requirements section (wraps ReqsSub subsections).

Constructors

ReqsProg [ReqsSub]

data ReqsSub where

Requirements subsections.

Constructors

FReqsSub :: Sentence -> [LabelledContent] -> ReqsSub
Functional requirements. LabelledContent for tables (includes input values).

FReqsSub' :: [LabelledContent] -> ReqsSub
Functional requirements. LabelledContent for tables (no input values).

NonFReqsSub :: ReqsSub
Non-Functional requirements.

data PDSub where

Problem Description subsections.

Constructors

TermsAndDefs :: Concept c => Maybe Sentence -> [c] -> PDSub
Terms and Definitions.

PhySysDesc :: Idea a => a -> [Sentence] -> LabelledContent -> [Contents] -> PDSub
Physical System Description.

Goals :: [Sentence] -> PDSub
Goals.

data ProblemDescription where

Problem Description section.

Constructors

PDProg :: Sentence -> [Section] -> [PDSub] -> ProblemDescription

newtype SSDSec

Specific System Description section (wraps SSDSub subsections).

Constructors

SSDProg [SSDSub]

data SSDSub where

Specific System Description subsections.

Constructors

SSDProblem :: ProblemDescription -> SSDSub
Problem description.

SSDSolChSpec :: SolChSpec -> SSDSub
Solution characteristics.

data SCSSub where

Solution Characteristics Specification subsections.

Constructors

Assumptions :: SCSSub
Assumptions.

TMs :: [Sentence] -> Fields -> SCSSub
Theory models.

GDs :: [Sentence] -> Fields -> DerivationDisplay -> SCSSub
General definitions.

DDs :: [Sentence] -> Fields -> DerivationDisplay -> SCSSub
Data definitions.

IMs :: [Sentence] -> Fields -> DerivationDisplay -> SCSSub
Instance models.

Constraints :: (HasUncertainty c, Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => Sentence -> [c] -> SCSSub
Constraints.

CorrSolnPpties :: (Quantity c, Constrained c) => [c] -> [Contents] -> SCSSub
Properties of a correct solution.

data SolChSpec where

Solution Characteristics Specification section (wraps SCSSub subsections).

Constructors

SCSProg :: [SCSSub] -> SolChSpec

Document subsection types needed for a SRS

Imported from Drasil.DocumentLanguage.Core

newtype AppndxSec

Appendix section.

Constructors

AppndxProg [Contents]

data AuxConstntSec

Values of Auxiliary Constants section.

Constructors

AuxConsProg CI [ConstQDef]

data DerivationDisplay

Choose whether to show or hide the derivation of an expression.

Constructors

ShowDerivation

HideDerivation

data Emphasis

How to handle emphasis of words.

Constructors

Bold

Italics

newtype OffShelfSolnsSec

Off-The-Shelf Solutions section.

Constructors

OffShelfSolnsProg [Contents]

newtype GSDSec

General System Description section (wraps GSDSub subsections).

Constructors

GSDProg [GSDSub]

data GSDSub where

General System Description subsections.

Constructors

SysCntxt :: [Contents] -> GSDSub
System context.

UsrChars :: [Contents] -> GSDSub
User characteristics.

SystCons :: [Contents] -> [Section] -> GSDSub
System constraints.

data IntroSec

Introduction section. Contents are top level followed by a list of subsections.

Constructors

IntroProg Sentence Sentence [IntroSub]
Temporary, will be modified once we've figured out more about the section.

data IntroSub where

Introduction subsections.

Constructors

IPurpose :: [Sentence] -> IntroSub
Describes purpose of the system.

IScope :: Sentence -> IntroSub
Describes scope of the system.

IChar :: [Sentence] -> [Sentence] -> [Sentence] -> IntroSub
Describes characteristics of the system.

IOrgSec :: CI -> Section -> Sentence -> IntroSub
Organises the section.

data LFunc where

Lens (lookup) functions (currently for TSymb).

Constructors

Term :: LFunc

Defn :: LFunc

TermExcept :: [DefinedQuantityDict] -> LFunc

DefnExcept :: [DefinedQuantityDict] -> LFunc

TAD

Fields

:: LFunc

Term and Definition.

data Literature

Types of literature.

Constructors

Lit Topic
Literature (with a Topic).

Doc' Topic
Existing documentation for (plural version of topic).

Manual Topic
Manual.

data RefSec

Reference section. Contents are top level followed by a list of subsections.

Constructors

RefProg Contents [RefTab]

data RefTab where

Reference subsections (tables) made out of units or symbols (can be customized).

Constructors

TUnits :: RefTab
Default table of units.

TUnits' :: [TUIntro] -> ([UnitDefn] -> LabelledContent) -> RefTab
Customized introduction.

TSymb :: [TSIntro] -> RefTab
Adds an introduction for a table of symbols.

TSymb' :: LFunc -> [TSIntro] -> RefTab
Allows Lens functions in addition to an introduction for a table of symbols.

TAandA :: [IdeaDict] -> RefTab
Default.

newtype StkhldrSec

Stakeholders section (wraps stakeholders subsections StkhldrSub).

Constructors

StkhldrProg [StkhldrSub]

data StkhldrSub where

Stakeholders subsections.

Constructors

Client :: CI -> Sentence -> StkhldrSub
May have a client.

Cstmr :: CI -> StkhldrSub
May have a customer.

data TConvention

Possible typographic conventions.

Constructors

Vector Emphasis
How vectors are emphasized.

Verb Sentence
Verbatim for specialized conventions.

newtype TraceabilitySec

Traceability Matices and Graphs section. Contains configurations (TraceConfig).

Constructors

TraceabilityProg [TraceConfig]

data TSIntro

For creating a table of symbols introduction

Constructors

TypogConvention [TConvention]
Typographic conventions used.

SymbOrder
Symbol ordering (defaults to alphabetical).

SymbConvention [Literature]
Symbol conventions match specified literature.

TSPurpose
Purpose of the Table of Symbols.

VectorUnits
Definition of vector components.

data TUIntro

For creating the table of units introduction.

Constructors

System
System of units (defaults to SI).

Derived
Sentence about derived units being used alongside SI.

TUPurpose
Purpose of the table of units.

Document subsection helper types

Imported from Drasil.DocumentLanguage.Definitions

data Field

Fields that should be displayed in definitions.

Constructors

Label

Symbol

Units

DefiningEquation

Description Verbosity InclUnits

Input

Output

InConstraints

OutConstraints

Notes

Source

RefBy

type Fields = [Field]

Synonym for a list of Fields.

data InclUnits

Determines whether to include or ignore units.

Constructors

IncludeUnits
In description field (for other symbols).

data Verbosity

Refers to the verbosity of statements.

Constructors

Verbose
Full Descriptions.

Succinct
Simple Description (do not redefine other symbols).

SRS Document creator functions

Imported from Drasil.DocumentLanguage

mkDoc :: SRSDecl -> (IdeaDict -> IdeaDict -> Sentence) -> System -> Document

Creates a document from a document description, a title combinator function, and system information.

fillcdbSRS :: SRSDecl -> System -> System

Assuming a given ChunkDB with no traces and minimal/no references, fill in for rest of system information. Currently fills in references, traceability matrix information and IdeaDicts.

Helper functions to make an SRS Document

Imports from various sections of drasil-docLang

intro :: Contents

Default Reference Material section introduction.

traceMatStandard :: System -> [TraceConfig]

Helpers to check if given argument has more than one peice of information

Contains traceability matrix assumptions, other assumptions, refinement, and other requirements.

tsymb :: [TSIntro] -> RefTab

Default is term and given introduction.

Table of symbols constructor.

tsymb'' :: [TSIntro] -> LFunc -> RefTab

Table of symbols constructor. Takes a custom function and introduction.

purpDoc :: CI -> Verbosity -> [Sentence]

Combines purpDocPara1 and developmentProcessParagraph. Verbosity controls if the developmentProcessParagraph is added or not.