Drasil.Sections.Introduction

Documentation

orgSec :: NamedIdea c => Sentence -> c -> Section -> Sentence -> Section

Doc.organization of the document section constructor. => Sentence -> c -> Section -> Sentence -> Section

introductionSection :: Sentence -> Sentence -> [Section] -> Section

Constructor for the introduction section problemIntroduction - Sentence introducing the specific example problem programDefinition - Sentence definition of the specific example subSections - List of subsections for this section

purposeOfDoc :: [Sentence] -> Section

constructor for purpose of document subsection purposeOfProgramParagraph - a sentence explaining the purpose of the specific example

scopeOfRequirements :: Sentence -> Section

constructor for scope of requirements subsection req - the main requirement for the program

charIntRdrF :: Idea a => a -> [Sentence] -> [Sentence] -> [Sentence] -> Section -> Section

constructor for characteristics of the intended reader subsection progName assumed topic asset r

purpDoc :: CI -> Verbosity -> [Sentence]