Drasil.Sections.Requirements

Description

Defines functions used in the Requirements section.

Requirements

reqF :: [Section] -> Section

Wrapper for reqIntro.

reqInputsRef :: Reference

Reference for the Required Inputs table.

Functional Requirements

fReqF :: [Contents] -> Section

Adds a generalized introduction for a Non-Fucntional Requirements section. Takes in the contents of that section.

Input Requirements

fullReqs :: (Quantity i, MayHaveUnit i) => [i] -> Sentence -> [ConceptInstance] -> [ConceptInstance]

Prepends a ConceptInstance referencing an input-value table to a list of other ConceptInstances. For listing input requirements.

fullTables :: (Quantity i, MayHaveUnit i) => [i] -> [LabelledContent] -> [LabelledContent]

Prepends given LabelledContent to an input-value table.

inReq :: Sentence -> ConceptInstance

Creates a ConceptInstance of input values.

inTable :: (Quantity i, MayHaveUnit i) => [i] -> LabelledContent

Creates a generalized input-value table for the Requirements section.

mkInputPropsTable :: (Quantity i, MayHaveUnit i, HasShortName r, Referable r) => [i] -> r -> LabelledContent

Creates an Input Data Table for use in the Functional Requirments section. Takes a list of wrapped variables and something that is Referable.

mkQRTuple :: (Quantity i, MayHaveUnit i, HasShortName i, Referable i) => [i] -> [(QuantityDict, Sentence)]

mkQRTupleRef :: (Quantity i, MayHaveUnit i, HasShortName r, Referable r) => [i] -> [r] -> [(QuantityDict, Sentence)]

mkValsSourceTable :: (Quantity i, MayHaveUnit i) => [(i, Sentence)] -> String -> Sentence -> LabelledContent

Creates a table for use in the Functional Requirments section. Takes a list of tuples containing variables and sources, a label, and a caption.

Non-functional Requirements

nfReqF :: [Contents] -> Section

Adds a generalized introduction for a Non-Fucntional Requirements section. Takes in the contents of that section.