Drasil.Sections.TableOfContents

Description

Standard code to make a table of contents.

Documentation

toToC :: DocSection -> ItemType

Finds all possible sections and subsections to make a Table of Contents.

findToC :: [DocSection] -> ShowTableOfContents

Finds whether the Table of Contents is in a SRSDecl.