Drasil.Sections.TableOfSymbols

Description

Standard code to make a table of symbols.

Documentation

table :: (Quantity s, MayHaveUnit s) => Stage -> [s] -> (s -> Sentence) -> LabelledContent

Table of Symbols creation function. Takes in a Stage, Symbols, and something that turns the symbols into a Sentence. Filters non-symbol chunks and checks for duplicate symbol error.

symbTableRef :: Reference

Makes a reference to the Table of Symbols.

tsymb :: [TSIntro] -> RefTab

Default is term and given introduction.

Table of symbols constructor.

tsymb' :: [TSIntro] -> RefTab

Similar to tsymb, but has a default Defn for the LFunc type. Still has a given introduction.

Table of symbols constructor.

tsymb'' :: [TSIntro] -> LFunc -> RefTab

Table of symbols constructor. Takes a custom function and introduction.

tsIntro :: [TSIntro] -> Contents

Table of symbols introduction builder. Used by mkRefSec.