Drasil.Sections.TraceabilityMandGs

Description

Defines functions used to create the Traceability Matrices and Graphs section.

Main Functions

generateTraceTable :: System -> LabelledContent

Makes a Traceability Table/Matrix that contains Items of Different Sections.

traceMatAssumpAssump :: TraceConfig

Assumptions on the assumptions of a traceability matrix.

traceMatAssumpOther :: TraceConfig

Other assumptions of the traceability matrix

traceMatRefinement :: TraceConfig

Refinement of the traceability matrix.

traceMatOtherReq :: System -> TraceConfig

Records other requirements. Converts the System into a TraceConfig.

traceMatStandard :: System -> [TraceConfig]

Helpers to check if given argument has more than one peice of information

Contains traceability matrix assumptions, other assumptions, refinement, and other requirements.

Helpers

tvAssumps :: TraceViewCat

Traceability viewing assumptions. Takes a UID and a ChunkDB. Returns a list of UIDs.

tvDataDefns :: TraceViewCat

Traceability viewing data definitions. Takes a UID and a ChunkDB. Returns a list of UIDs.

tvGenDefns :: TraceViewCat

Traceability viewing general definitions. Takes a UID and a ChunkDB. Returns a list of UIDs.

tvTheoryModels :: TraceViewCat

Traceability viewing theory models. Takes a UID and a ChunkDB. Returns a list of UIDs.

tvInsModels :: TraceViewCat

Traceability viewing instance models. Takes a UID and a ChunkDB. Returns a list of UIDs.

tvGoals :: TraceViewCat

Traceability viewing goals. Takes a UID and a ChunkDB. Returns a list of UIDs.

tvReqs :: TraceViewCat

Traceability viewing requirements. Takes a UID and a ChunkDB. Returns a list of UIDs.

tvChanges :: TraceViewCat

Traceability viewing changes. Takes a UID and a ChunkDB. Returns a list of UIDs.