Drasil.DocLang

Documentation

newtype AppndxSec

Constructors

AppndxProg [Contents]

data AuxConstntSec

Values of Auxiliary Constants section

Constructors

AuxConsProg CI [QDefinition]

data DerivationDisplay

Constructors

ShowDerivation

HideDerivation

type DocDesc = [DocSection]

data DocSection

Document sections are either Reference, Introduction, or Specific System Description sections (for now!)

Constructors

RefSec RefSec

IntroSec IntroSec

StkhldrSec StkhldrSec

GSDSec GSDSec

SSDSec SSDSec

ReqrmntSec ReqrmntSec

LCsSec LCsSec

LCsSec' LCsSec'

UCsSec UCsSec

TraceabilitySec TraceabilitySec

AuxConstntSec AuxConstntSec

Bibliography

AppndxSec AppndxSec

ExistingSolnSec ExistingSolnSec

data Emphasis

How to handle Emphasis

Constructors

Bold

Italics

data GSDSec

Constructors

GSDProg2 [GSDSub]

data GSDSub where

Constructors

SysCntxt :: [Contents] -> GSDSub

UsrChars :: [Contents] -> GSDSub

SystCons :: [Contents] -> [Section] -> GSDSub

data IntroSec

Introduction section. Contents are top level followed by a list of subsections.

Constructors

IntroProg Sentence Sentence [IntroSub]
Temporary, will be modified once we've figured out more about the section.

data IntroSub where

Introduction subsections

Constructors

IPurpose :: Sentence -> IntroSub

IScope :: Sentence -> Sentence -> IntroSub

IChar :: [Sentence] -> [Sentence] -> [Sentence] -> IntroSub

IOrgSec :: Sentence -> CI -> Section -> Sentence -> IntroSub

newtype LCsSec

Constructors

LCsProg [Contents]

data LFunc where

Lens (lookup) functions (currently for TSymb)

Constructors

Term :: LFunc

Defn :: LFunc

TermExcept :: [DefinedQuantityDict] -> LFunc

DefnExcept :: [DefinedQuantityDict] -> LFunc

TAD :: LFunc

data Literature

Types of literature

Constructors

Lit Topic
literature

Doc' Topic
existing documentation for (plural of topic)

Manual Topic
manual

data ProblemDescription where

Problem Description section

Constructors

PDProg :: Sentence -> CI -> Sentence -> [Section] -> ProblemDescription

data RefSec

Reference section. Contents are top level followed by a list of subsections.

Constructors

RefProg Contents [RefTab]

data RefTab where

Reference subsections

Constructors

TUnits :: RefTab

TUnits' :: [TUIntro] -> RefTab

TSymb :: [TSIntro] -> RefTab

TSymb' :: LFunc -> [TSIntro] -> RefTab

TAandA :: RefTab

newtype ReqrmntSec

Constructors

ReqsProg [ReqsSub]

data ReqsSub where

Constructors

FReqsSub :: [ConceptInstance] -> [LabelledContent] -> ReqsSub

NonFReqsSub :: [ConceptInstance] -> ReqsSub

data SCSSub where

Solution Characteristics Specification subsections

Constructors

Assumptions :: SCSSub

TMs :: [Sentence] -> Fields -> [TheoryModel] -> SCSSub

GDs :: [Sentence] -> Fields -> [GenDefn] -> DerivationDisplay -> SCSSub

DDs :: [Sentence] -> Fields -> [DataDefinition] -> DerivationDisplay -> SCSSub

IMs :: [Sentence] -> Fields -> [InstanceModel] -> DerivationDisplay -> SCSSub

Constraints :: Sentence -> Sentence -> Sentence -> [LabelledContent] -> SCSSub

CorrSolnPpties :: [Contents] -> SCSSub

newtype SSDSec

Specific System Description section . Contains a list of subsections.

Constructors

SSDProg [SSDSub]

data SSDSub where

Specific system description subsections

Constructors

SSDSubVerb :: Section -> SSDSub

SSDProblem :: ProblemDescription -> SSDSub

SSDSolChSpec :: SolChSpec -> SSDSub

data SolChSpec where

Solution Characteristics Specification section

Constructors

SCSProg :: [SCSSub] -> SolChSpec

newtype ExistingSolnSec

Off-The-Shelf Solutions section

Constructors

ExistSolnProg [Contents]

data StkhldrSec

Stakeholders section

Constructors

StkhldrProg2 [StkhldrSub]

newtype LCsSec'

Constructors

LCsProg' [ConceptInstance]

data StkhldrSub where

Stakeholders subsections

Constructors

Client :: CI -> Sentence -> StkhldrSub

Cstmr :: CI -> StkhldrSub

data TConvention

Possible typographic conventions

Constructors

Vector Emphasis
How vectors are emphasized

Verb Sentence
Verbatim for specialized conventions

data TraceabilitySec

Constructors

TraceabilityProg [LabelledContent] [Sentence] [Contents] [Section]

data TSIntro

For creating the table of symbols intro

Constructors

TypogConvention [TConvention]
Typographic conventions used

SymbOrder
Symbol ordering (defaults to alphabetical)

SymbConvention [Literature]
Symbol conventions match specified literature

TSPurpose
Purpose of the Table of Symbols

newtype UCsSec

Constructors

UCsProg [Contents]

mkDoc :: DocDesc -> (IdeaDict -> IdeaDict -> Sentence) -> SystemInformation -> Document

Creates a document from a document description and system information

tsymb :: [TSIntro] -> RefTab

Default Term and given intro

table of symbols constructor

tsymb'' :: [TSIntro] -> LFunc -> RefTab

Custom function and intro.

Custom table of symbols constructor

mkEnumSimple :: (a -> ListTuple) -> [a] -> [Contents]

mkEnumSimple is a convenience function for converting lists into Simple-type Enumerations.

mkEnumSimpleD :: (Referable c, HasShortName c, Definition c) => [c] -> [Contents]

mkEnumSimpleD is a convenience function for transforming types which are instances of the constraints Referable, HasShortName, and Definition, into Simple-type Enumerations.

data Field

Fields that should be displayed in definitions

Constructors

Label

Symbol

Units

DefiningEquation

Description Verbosity InclUnits

Input

Output

InConstraints

OutConstraints

Notes

Source

RefBy

type Fields = [Field]

Synonym for a list of Field

data InclUnits

Constructors

IncludeUnits

data Verbosity

Constructors

Verbose

ddefn :: Fields -> SystemInformation -> DataDefinition -> LabelledContent

Create a data definition using a list of fields, a database of symbols, and a QDefinition (called automatically by SCSSub program)

data ModelDB

ddRefDB :: ModelDB -> RefMap DataDefinition

mdb :: [TheoryModel] -> [GenDefn] -> [DataDefinition] -> [InstanceModel] -> ModelDB

valsOfAuxConstantsF :: Idea a => a -> [QDefinition] -> Section

genSysF :: [Section] -> Contents -> [Contents] -> [Section] -> Section

intro :: Contents

Default reference section introduction

reqF :: [Section] -> Section

mkInputPropsTable :: (Quantity i, MayHaveUnit i, HasShortName r, Referable r) => [i] -> r -> LabelledContent

takes a list of wrapped variables and creates an Input Data Table for uses in Functional Requirments

mkQRTuple :: (Quantity i, MayHaveUnit i, HasShortName i, Referable i) => [i] -> [(QuantityDict, Sentence)]

mkQRTupleRef :: (Quantity i, MayHaveUnit i, HasShortName r, Referable r) => [i] -> [r] -> [(QuantityDict, Sentence)]

mkValsSourceTable :: (Quantity i, MayHaveUnit i) => [(i, Sentence)] -> String -> Sentence -> LabelledContent

takes a list of tuples of variables and sources and creates an table for uses in Functional Requirments

assumpF :: Section -> Section -> Section -> Section -> Section -> Section -> [Contents] -> Section

dataConstraintUncertainty :: Sentence

dataDefnF :: Sentence -> [Contents] -> Section

goalStmtF :: [Sentence] -> [Contents] -> Section

inDataConstTbl :: (HasUncertainty c, Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => [c] -> LabelledContent

inModelF :: Section -> Section -> Section -> Section -> [Contents] -> Section

outDataConstTbl :: (Quantity c, Constrained c) => [c] -> LabelledContent

physSystDesc :: Sentence -> LabelledContent -> [Contents] -> Section

probDescF :: Idea a => Sentence -> a -> Sentence -> [Section] -> Section

termDefnF :: Maybe Sentence -> [Contents] -> Section

specSysDescr :: [Section] -> Section

Specific System description section builder. Takes the system and subsections.

traceGIntro :: [LabelledContent] -> [Sentence] -> [UnlabelledContent]

traceMGF :: [LabelledContent] -> [Sentence] -> [Contents] -> [Section] -> Section

generateTraceTable :: SystemInformation -> LabelledContent

getDocDesc :: DocDesc -> [Sentence]

egetDocDesc :: DocDesc -> [Expr]

ciGetDocDesc :: DocDesc -> [CI]

generateTraceMap :: [DocSection] -> TraceMap

getTraceMapFromTM :: [SCSSub] -> [TheoryModel]

getTraceMapFromGD :: [SCSSub] -> [GenDefn]

getTraceMapFromDD :: [SCSSub] -> [DataDefinition]

getTraceMapFromIM :: [SCSSub] -> [InstanceModel]

getSCSSub :: [DocSection] -> [SCSSub]

generateTraceMap' :: [ConceptInstance] -> TraceMap

goalStmtLabel :: Reference

solutionLabel :: Reference

characteristicsLabel :: Reference

physSystDescriptionLabel :: Reference