Drasil.DocLang.Notebook

Section Constructors

intro :: [Contents] -> [Section] -> Section

Section constructors for the lesson plan documents/jupyter notebooks. should be fixed once we have a more concrete idea of the notebook structure TODO: Remove [Section] since the structure of lesson plans should not be nested maybe add a new type Chapter for lesson plans?

learnObj :: [Contents] -> [Section] -> Section

Section constructors for the lesson plan documents/jupyter notebooks. should be fixed once we have a more concrete idea of the notebook structure TODO: Remove [Section] since the structure of lesson plans should not be nested maybe add a new type Chapter for lesson plans?

review :: [Contents] -> [Section] -> Section

Section constructors for the lesson plan documents/jupyter notebooks. should be fixed once we have a more concrete idea of the notebook structure TODO: Remove [Section] since the structure of lesson plans should not be nested maybe add a new type Chapter for lesson plans?

caseProb :: [Contents] -> [Section] -> Section

Section constructors for the lesson plan documents/jupyter notebooks. should be fixed once we have a more concrete idea of the notebook structure TODO: Remove [Section] since the structure of lesson plans should not be nested maybe add a new type Chapter for lesson plans?

summary :: [Contents] -> [Section] -> Section

Section constructors for the lesson plan documents/jupyter notebooks. should be fixed once we have a more concrete idea of the notebook structure TODO: Remove [Section] since the structure of lesson plans should not be nested maybe add a new type Chapter for lesson plans?

appendix :: [Contents] -> [Section] -> Section

Section constructors for the lesson plan documents/jupyter notebooks. should be fixed once we have a more concrete idea of the notebook structure TODO: Remove [Section] since the structure of lesson plans should not be nested maybe add a new type Chapter for lesson plans?

reference :: [Contents] -> [Section] -> Section

Section constructors for the lesson plan documents/jupyter notebooks. should be fixed once we have a more concrete idea of the notebook structure TODO: Remove [Section] since the structure of lesson plans should not be nested maybe add a new type Chapter for lesson plans?

example :: [Contents] -> [Section] -> Section

Section constructors for the lesson plan documents/jupyter notebooks. should be fixed once we have a more concrete idea of the notebook structure TODO: Remove [Section] since the structure of lesson plans should not be nested maybe add a new type Chapter for lesson plans?

sectionReferences :: [Reference]

Section References

introLabel :: Reference

Individual section reference labels. Used in creating example sections for the notebook.

learnObjLabel :: Reference

Individual section reference labels. Used in creating example sections for the notebook.

docPurposeLabel :: Reference

Individual section reference labels. Used in creating example sections for the notebook.

referenceLabel :: Reference

Individual section reference labels. Used in creating example sections for the notebook.

reviewLabel :: Reference

Individual section reference labels. Used in creating example sections for the notebook.

caseProbLabel :: Reference

Individual section reference labels. Used in creating example sections for the notebook.

appendixLabel :: Reference

Individual section reference labels. Used in creating example sections for the notebook.

summaryLabel :: Reference

Individual section reference labels. Used in creating example sections for the notebook.

exampleLabel :: Reference

Individual section reference labels. Used in creating example sections for the notebook.