Drasil.DocumentLanguage

Description

Start the process of moving away from Document as the main internal representation of information, to something more informative. Over time, we'll want to have a cleaner separation, but doing that all at once would break too much for too long. So we start here instead.

Documentation

mkDoc :: SRSDecl -> (IdeaDict -> IdeaDict -> Sentence) -> (System, DocDesc) -> Document

Creates a document from a document description, a title combinator function, and system information.

fillcdbSRS :: SRSDecl -> System -> (System, DocDesc)

Assuming a given ChunkDB with no traces and minimal/no references, fill in for rest of system information. Currently fills in references, traceability matrix information and IdeaDicts.

findAllRefs :: Section -> [Reference]

Recursively find all references in a section (meant for getting at LabelledContent).