Drasil.DocumentLanguage

Description

Start the process of moving away from Document as the main internal representation of information, to something more informative. Over time, we'll want to have a cleaner separation, but doing that all at once would break too much for too long. So we start here instead.

Documentation

mkDoc :: System -> SRSDecl -> (IdeaDict -> CI -> Sentence) -> (Document, System)

Creates a document from a System, a document description (SRSDecl), and a title combinator.

findAllRefs :: Section -> [Reference]

Recursively find all references in a section (meant for getting at LabelledContent).