Drasil.DocumentLanguage

Description

Start the process of moving away from Document as the main internal representation of information, to something more informative. Over time, we'll want to have a cleaner separation, but doing that all at once would break too much for too long. So we start here instead.

Documentation

mkDoc :: SRSDecl -> (IdeaDict -> IdeaDict -> Sentence) -> SystemInformation -> Document

Creates a document from a document description, a title combinator function, and system information.

fillTraceSI :: SRSDecl -> SystemInformation -> SystemInformation

Helper for filling in the traceability matrix and graph information into the system.

fillTraceMaps :: DocDesc -> SystemInformation -> SystemInformation

Fills in the traceabiliy matrix and graphs section of the system information using the document description.

fillReqs :: DocDesc -> SystemInformation -> SystemInformation

Fills in the requirements section of the system information using the document description.

extractUnits :: DocDesc -> ChunkDB -> [UnitDefn]

Constructs the unit definitions (UnitDefns) found in the document description (DocDesc) from a database (ChunkDB).

findToC :: [DocSection] -> ShowTableOfContents

Finds whether the Table of Contents is in a SRSDecl.

mkSections :: SystemInformation -> DocDesc -> [Section]

Helper for creating the different document sections.

mkToC :: DocDesc -> Section

Helper for making the Table of Contents section.

mkRefSec :: SystemInformation -> DocDesc -> RefSec -> Section

Helper for creating the reference section and subsections. Includes Table of Symbols, Units and Abbreviations and Acronyms.

tunit :: [TUIntro] -> RefTab

Table of units with an SI Name.

Table of units constructors.

tunit' :: [TUIntro] -> RefTab

Table of units with SI name in the description column.

Table of units constructors.

mkTSymb :: (Quantity e, Concept e, Eq e, MayHaveUnit e) => [e] -> LFunc -> [TSIntro] -> Section

Helper for creating the table of symbols.

tsymb :: [TSIntro] -> RefTab

Default is term and given introduction.

Table of symbols constructor.

tsymb' :: [TSIntro] -> RefTab

Similar to tsymb, but has a default Defn for the LFunc type. Still has a given introduction.

Table of symbols constructor.

tsymb'' :: [TSIntro] -> LFunc -> RefTab

Table of symbols constructor. Takes a custom function and introduction.

tsIntro :: [TSIntro] -> Contents

Table of symbols introduction builder. Used by mkRefSec.

tsI :: TSIntro -> Sentence

Table of symbols intro writer. Translates a TSIntro to a list in a Sentence.

typogConvention :: [TConvention] -> Sentence

Typographic convention writer. Translates a list of typographic conventions (TConventions) to a Sentence.

symbConvention :: [Literature] -> Sentence

Symbolic convention writer.

tuIntro :: [TUIntro] -> Contents

Table of units introduction builder. Used by mkRefSec.

tuI :: TUIntro -> Sentence

Table of units introduction writer. Translates a TUIntro to a Sentence.

defaultTUI :: [TUIntro]

Default table of units intro that contains the system, derivation, and purpose.

mkIntroSec :: SystemInformation -> IntroSec -> Section

Makes the Introduction section into a Section.

mkStkhldrSec :: StkhldrSec -> Section

Helper for making the Stakeholders section.

mkGSDSec :: GSDSec -> Section

Helper for making the General System Description section.

mkSSDSec :: SystemInformation -> SSDSec -> Section

Helper for making the Specific System Description section.

mkSSDProb :: SystemInformation -> ProblemDescription -> Section

Helper for making the Specific System Description Problem section.

mkSolChSpec :: SystemInformation -> SolChSpec -> Section

Helper for making the Solution Characteristics Specification section.

helperCI :: ConceptInstance -> SystemInformation -> ConceptInstance

Helper for making a ConceptInstance with a reference to the system information.

tmStub :: Section

Section stubs for implicit referencing of different models and definitions.

ddStub :: Section

Section stubs for implicit referencing of different models and definitions.

imStub :: Section

Section stubs for implicit referencing of different models and definitions.

pdStub :: Section

Section stubs for implicit referencing of different models and definitions.

mkReqrmntSec :: ReqrmntSec -> Section

Helper for making the Requirements section.

mkLCsSec :: LCsSec -> Section

Helper for making the Likely Changes section.

mkUCsSec :: UCsSec -> Section

Helper for making the Unikely Changes section.

mkTraceabilitySec :: TraceabilitySec -> SystemInformation -> Section

Helper for making the Traceability Matrices and Graphs section.

mkOffShelfSolnSec :: OffShelfSolnsSec -> Section

Helper for making the Off-the-Shelf Solutions section.

mkAuxConsSec :: AuxConstntSec -> Section

Helper for making the Values of Auxiliary Constants section.

mkBib :: BibRef -> Section

Helper for making the Bibliography section.

mkAppndxSec :: AppndxSec -> Section

Helper for making the Appendix section.

siSys :: SystemInformation -> IdeaDict

Helper to get the program name as an IdeaDict.