Drasil.DocumentLanguage

Description

Start the process of moving away from Document as the main internal representation of information, to something more informative. Over time, we'll want to have a cleaner separation, but doing that all at once would break too much for too long. So we start here instead.

Main Function

mkDoc :: SRSDecl -> (IdeaDict -> IdeaDict -> Sentence) -> System -> Document

Creates a document from a document description, a title combinator function, and system information.

Functions to Fill ChunkDB

fillcdbSRS :: SRSDecl -> System -> System

Assuming a given ChunkDB with no traces and minimal/no references, fill in for rest of system information. Currently fills in references, traceability matrix information and IdeaDicts.

fillSecAndLC :: SRSDecl -> System -> System

Fill in the Sections and LabelledContent maps of the ChunkDB from the SRSDecl.

fillReferences :: SRSDecl -> System -> System

Takes in existing information from the Chunk database to construct a database of references.

dRefToRef :: HasDecRef a => a -> [Reference]

Helper that gets references from definitions and models.

findAllRefs :: Section -> [Reference]

Recursively find all references in a section (meant for getting at LabelledContent).

fillTraceSI :: SRSDecl -> System -> System

Helper for filling in the traceability matrix and graph information into the system.

fillTraceMaps :: DocDesc -> System -> System

Fills in the traceabiliy matrix and graphs section of the system information using the document description.

fillReqs :: DocDesc -> System -> System

Fills in the requirements section of the system information using the document description.

extractUnits :: DocDesc -> ChunkDB -> [UnitDefn]

Constructs the unit definitions (UnitDefns) found in the document description (DocDesc) from a database (ChunkDB).

Section Creator Functions

mkSections :: System -> DocDesc -> [Section]

Helper for creating the different document sections.

Table of Contents

mkToC :: DocDesc -> Section

Helper for making the Table of Contents section.

Reference Materials

mkRefSec :: System -> DocDesc -> RefSec -> Section

Helper for creating the reference section and subsections. Includes Table of Symbols, Units and Abbreviations and Acronyms.

mkTSymb :: (Quantity e, Concept e, Eq e, MayHaveUnit e) => [e] -> LFunc -> [TSIntro] -> Section

Helper for creating the table of symbols.

Introduction

mkIntroSec :: System -> IntroSec -> Section

Makes the Introduction section into a Section.

Stakeholders

mkStkhldrSec :: StkhldrSec -> Section

Helper for making the Stakeholders section.

General System Description

mkGSDSec :: GSDSec -> Section

Helper for making the General System Description section.

Specific System Description

mkSSDSec :: System -> SSDSec -> Section

Helper for making the Specific System Description section.

mkSSDProb :: System -> ProblemDescription -> Section

Helper for making the Specific System Description Problem section.

mkSolChSpec :: System -> SolChSpec -> Section

Helper for making the Solution Characteristics Specification section.

Requirements

mkReqrmntSec :: ReqrmntSec -> Section

Helper for making the Requirements section.

Likely Changes

mkLCsSec :: LCsSec -> Section

Helper for making the Likely Changes section.

Unlikely Changes

mkUCsSec :: UCsSec -> Section

Helper for making the Unikely Changes section.

introChgs :: NamedIdea n => n -> [ConceptInstance] -> Contents

Intro paragraph for likely and unlikely changes

Traceability

mkTraceabilitySec :: TraceabilitySec -> System -> Section

Helper for making the Traceability Matrices and Graphs section.

header :: ([UID] -> [UID]) -> System -> [Sentence]

Helper to get headers of rows and columns

Off the Shelf Solutions

mkOffShelfSolnSec :: OffShelfSolnsSec -> Section

Helper for making the Off-the-Shelf Solutions section.

Auxiliary Constants

mkAuxConsSec :: AuxConstntSec -> Section

Helper for making the Values of Auxiliary Constants section.

References

mkBib :: BibRef -> Section

Helper for making the References section.

Appendix

mkAppndxSec :: AppndxSec -> Section

Helper for making the Appendix section.