Drasil.DocumentLanguage.TraceabilityGraph

Description

Defines functions to create traceability graphs in SRS documents.

Main Functions

traceMGF :: [LabelledContent] -> [Sentence] -> [Contents] -> String -> [Section] -> Section

Wrapper for traceMIntro and traceGIntro. Turns references (LabelledContents), trailing notes (Sentences), and any other needed contents to create a Traceability Section. Traceability graphs generate as both a link and a figure for convenience.

traceGIntro :: [UID] -> [Sentence] -> [UnlabelledContent]

Generalized traceability graph introduction: appends references to the traceability graphs in Sentence form and wraps in Contents. Usually references the five graphs as defined in GraphInfo.

mkGraphInfo :: System -> GraphInfo

Extracts traceability graph inforomation from filled-in System.

Helper Functions

mkGraphNodes :: TraceViewCat -> System -> String -> NodeFamily

Gets the node family of a graph based on the given section and system information. Also applies a given colour to the node family.

mkGraphEdges :: [TraceViewCat] -> [TraceViewCat] -> System -> [(UID, [UID])]

Creates the graph edges based on the relation of the first list of sections to the second. Also needs the system information. Return value is of the form (Section, [Dependencies]).

makeTGraph :: [UID] -> [[UID]] -> [UID] -> [(UID, [UID])]

Helper for making graph edges. Taken from Utils.Drasil's traceability matrix relation finder. But, instead of marking X on two related ideas, it makes them an edge.

checkUID :: UID -> System -> UID

Checker for uids by finding if the UID is in one of the possible data sets contained in the System database.

checkUIDAbbrev :: System -> UID -> String

Similar to checkUID but prepends domain for labelling.

checkUIDRefAdd :: System -> UID -> String

Similar to checkUID but gets reference addresses for display.

traceGHeader :: (ChunkDB -> [UID]) -> System -> [UID]

Helper that finds the header of a traceability matrix. However, here we use this to get a list of UIDs for a traceability graph instead.

traceGRowHeader :: ([UID] -> [UID]) -> System -> [UID]

Helper that finds the headers of the traceability matrix rows. However, here we use this to get a list of UIDs for a traceability graph instead. This is then used to create the graph edges.

graphShows :: UID -> Sentence -> Sentence

Helper that makes references of the form "reference shows the dependencies of something". Only takes a reference UID instead of a Reference.

Functions to Create a Traceability Graphs

Functions related to setting up the structure and contents of the traceability graphs section.

allvsallDesc :: Sentence

Description of the AllvsAll traceability graph.

traceGLst :: Contents

Create a list of traceability graph references.

traceGCon :: String -> [Contents]

The Traceability Graph contents.

traceGraphLC :: String -> FilePath -> UID -> LabelledContent

Generates traceability graphs as figures on an SRS document.

traceGFiles :: [String]

Traceability graph file names.

traceGUIDs :: [UID]

Traceabiliy graph reference UIDs.

traceyGraphPaths :: String -> [String]

Create reference paths to traceability graphs given an example name. For .pdf links

traceyGraphGetRefs :: String -> [Reference]

Create references to traceability graphs given an example name. Primarily used for reference database in examples.

traceyGraphPath :: String -> String -> String

Gets the path of a traceability graph given an example folder name and the graph name. For .png files

resourcePath :: String

Traceability graphs reference path.

folderList' :: [ItemType]

Helper to create a list of traceability graph references.