Drasil.Sections.SpecificSystemDescription

Description

Defines functions used to create the Specific System Description section.

Specific System Description

specSysDescr :: [Section] -> Section

Specific System Description section builder.

Problem Description

probDescF :: Sentence -> [Section] -> Section

Describes a problem the system is needed to accomplish.

termDefnF :: Concept c => Maybe Sentence -> [c] -> Section

Creates the Terms and Definitions section. Can take a (Just Sentence) if needed or Nothing if not. Also takes Concepts that contain the definitions.

termDefnF' :: Maybe Sentence -> [Contents] -> Section

Similar to termDefnF, except does not take definitions from the list of terms.

physSystDesc :: Idea a => a -> [Sentence] -> LabelledContent -> [Contents] -> Section

General introduction for the Physical System Description section.

goalStmtF :: [Sentence] -> [Contents] -> Int -> Section

General constructor for the Goal Statement section. Takes the given inputs (Sentences) and the descriptions (Contents).

Solution Characteristics Specification

solutionCharSpecIntro :: Idea a => a -> Section -> Contents

General introduction for the Solution Characteristics Specification section. Takes the program name and a section of instance models.

assumpF :: [Contents] -> Section

Creates an Assumptions section by prepending a general introduction to other related Contents.

thModF :: Idea a => a -> [Contents] -> Section

Wrapper for thModelIntro. Takes the program name and other Contents.

genDefnF :: [Contents] -> Section

Creates a General Definitions section with a general introduction. Takes in relevant general definitions (Contents). Use empty list if none are needed.

dataDefnF :: Sentence -> [Contents] -> Section

Similar to genDefnF, but for Data Definitions. It also uses EmptyS if the ending Sentence is not needed rather than an empty list.

inModelF :: Section -> Section -> Section -> Section -> [Contents] -> Section

Constructor for Instance Models. Takes the problem description, data definition, theoretical model, general definition, and any other relevant contents.

datConF :: (HasUncertainty c, Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => Sentence -> [c] -> Section

Constructor for Data Constraints section. Takes a trailing Sentence (use EmptyS if none) and data constraints.

inDataConstTbl :: (HasUncertainty c, Quantity c, Constrained c, HasReasVal c, MayHaveUnit c) => [c] -> LabelledContent

Creates the input Data Constraints Table.

outDataConstTbl :: (Quantity c, Constrained c) => [c] -> LabelledContent

Creates the output Data Constraints Table.

propCorSolF :: (Quantity c, Constrained c) => [c] -> [Contents] -> Section

Creates the Properties of a Correct Solution section.

auxSpecSent :: Sentence

General Sentence that describes some auxiliary specifications of the system.

tInDataCstRef :: Reference

Input/Output Data Constraint Table references.

tOutDataCstRef :: Reference

Input/Output Data Constraint Table references.

helperCI :: ConceptInstance -> System -> ConceptInstance

Helper for making a ConceptInstance with a reference to the system information. Used to find where a particular assumption is referenced.

Subsection Stubs

tmStub :: Section

Section stubs for implicit referencing of different models and definitions.

ddStub :: Section

Section stubs for implicit referencing of different models and definitions.

gdStub :: Section

Section stubs for implicit referencing of different models and definitions.

imStub :: Section

Section stubs for implicit referencing of different models and definitions.

pdStub :: Section

Section stubs for implicit referencing of different models and definitions.